@@ -1,68 +1,106 @@
 {
-  "_from": "bson@^1.1.4",
-  "_id": "bson@1.1.6",
+  "_from": "bson@^4.2.2",
+  "_id": "bson@4.5.3",
   "_inBundle": false,
-  "_integrity": "sha512-EvVNVeGo4tHxwi8L6bPj3y3itEvStdwvvlojVxxbyYfoaxJ6keLgrTuKdyfEAszFK+H3olzBuafE0yoh0D1gdg==",
+  "_integrity": "sha512-qVX7LX79Mtj7B3NPLzCfBiCP6RAsjiV8N63DjlaVVpZW+PFoDTxQ4SeDbSpcqgE6mXksM5CAwZnXxxxn/XwC0g==",
   "_location": "/bson",
-  "_phantomChildren": {},
+  "_phantomChildren": {
+    "base64-js": "1.5.1",
+    "ieee754": "1.2.1"
+  },
   "_requested": {
     "type": "range",
     "registry": true,
-    "raw": "bson@^1.1.4",
+    "raw": "bson@^4.2.2",
     "name": "bson",
     "escapedName": "bson",
-    "rawSpec": "^1.1.4",
+    "rawSpec": "^4.2.2",
     "saveSpec": null,
-    "fetchSpec": "^1.1.4"
+    "fetchSpec": "^4.2.2"
   },
   "_requiredBy": [
     "/mongodb",
     "/mongoose"
   ],
-  "_resolved": "https://registry.npmjs.org/bson/-/bson-1.1.6.tgz",
-  "_shasum": "fb819be9a60cd677e0853aee4ca712a785d6618a",
-  "_spec": "bson@^1.1.4",
-  "_where": "D:\\Apache\\Apache24\\htdocs\\Server\\React\\DemoProject\\demo\\node_modules\\mongoose",
+  "_resolved": "https://registry.npmjs.org/bson/-/bson-4.5.3.tgz",
+  "_shasum": "de3783b357a407d935510beb1fbb285fef43bb06",
+  "_spec": "bson@^4.2.2",
+  "_where": "D:\\Apache\\Apache24\\htdocs\\Server\\React\\DemoProject\\frontend\\node_modules\\mongoose",
   "author": {
-    "name": "Christian Amor Kvalheim",
-    "email": "christkv@gmail.com"
+    "name": "The MongoDB NodeJS Team",
+    "email": "dbx-node@mongodb.com"
   },
-  "browser": "lib/bson/bson.js",
+  "browser": {
+    "./lib/bson.js": "./dist/bson.browser.umd.js",
+    "./dist/bson.esm.js": "./dist/bson.browser.esm.js"
+  },
   "bugs": {
-    "url": "https://github.com/mongodb/js-bson/issues"
+    "url": "https://jira.mongodb.org/projects/NODE/issues/"
   },
   "bundleDependencies": false,
   "config": {
     "native": false
   },
   "contributors": [],
+  "dependencies": {
+    "buffer": "^5.6.0"
+  },
   "deprecated": false,
   "description": "A bson parser for node.js and the browser",
   "devDependencies": {
-    "babel-core": "^6.14.0",
-    "babel-loader": "^6.2.5",
-    "babel-polyfill": "^6.13.0",
-    "babel-preset-es2015": "^6.14.0",
-    "babel-preset-stage-0": "^6.5.0",
-    "babel-register": "^6.14.0",
-    "benchmark": "1.0.0",
-    "colors": "1.1.0",
-    "conventional-changelog-cli": "^1.3.5",
-    "nodeunit": "0.9.0",
-    "standard-version": "^9.1.1",
-    "webpack": "^1.13.2",
-    "webpack-polyfills-plugin": "0.0.9"
-  },
-  "directories": {
-    "lib": "./lib/bson"
+    "@babel/plugin-external-helpers": "^7.10.4",
+    "@babel/preset-env": "^7.11.0",
+    "@istanbuljs/nyc-config-typescript": "^1.0.1",
+    "@microsoft/api-extractor": "^7.11.2",
+    "@rollup/plugin-babel": "^5.2.0",
+    "@rollup/plugin-commonjs": "^15.0.0",
+    "@rollup/plugin-json": "^4.1.0",
+    "@rollup/plugin-node-resolve": "^9.0.0",
+    "@rollup/plugin-typescript": "^6.0.0",
+    "@typescript-eslint/eslint-plugin": "^3.10.1",
+    "@typescript-eslint/parser": "^3.10.1",
+    "array-includes": "^3.1.3",
+    "benchmark": "^2.1.4",
+    "chai": "^4.2.0",
+    "downlevel-dts": "^0.7.0",
+    "eslint": "^7.7.0",
+    "eslint-config-prettier": "^6.11.0",
+    "eslint-plugin-prettier": "^3.1.4",
+    "eslint-plugin-tsdoc": "^0.2.6",
+    "karma": "^6.3.4",
+    "karma-chai": "^0.1.0",
+    "karma-chrome-launcher": "^3.1.0",
+    "karma-mocha": "^2.0.1",
+    "karma-mocha-reporter": "^2.2.5",
+    "karma-rollup-preprocessor": "^7.0.5",
+    "mocha": "5.2.0",
+    "node-fetch": "^2.6.1",
+    "nyc": "^15.1.0",
+    "object.entries": "^1.1.4",
+    "prettier": "^2.1.1",
+    "rimraf": "^3.0.2",
+    "rollup": "^2.26.5",
+    "rollup-plugin-commonjs": "^10.1.0",
+    "rollup-plugin-node-globals": "^1.4.0",
+    "rollup-plugin-node-polyfills": "^0.2.1",
+    "rollup-plugin-polyfill-node": "^0.7.0",
+    "standard-version": "^9.3.0",
+    "ts-node": "^9.0.0",
+    "tsd": "^0.17.0",
+    "typedoc": "^0.21.2",
+    "typescript": "^4.0.2",
+    "typescript-cached-transpile": "0.0.6",
+    "uuid": "^8.3.2"
   },
   "engines": {
-    "node": ">=0.6.19"
+    "node": ">=6.9.0"
   },
   "files": [
     "lib",
-    "index.js",
-    "browser_build",
+    "src",
+    "dist",
+    "bson.d.ts",
+    "etc/prepare.js",
     "bower.json"
   ],
   "homepage": "https://github.com/mongodb/js-bson#readme",
