@@ -1,82 +1,99 @@
 {
-  "_from": "mongodb@3.6.8",
-  "_id": "mongodb@3.6.8",
+  "_from": "mongodb@4.1.2",
+  "_id": "mongodb@4.1.2",
   "_inBundle": false,
-  "_integrity": "sha512-sDjJvI73WjON1vapcbyBD3Ao9/VN3TKYY8/QX9EPbs22KaCSrQ5rXo5ZZd44tWJ3wl3FlnrFZ+KyUtNH6+1ZPQ==",
+  "_integrity": "sha512-pHCKDoOy1h6mVurziJmXmTMPatYWOx8pbnyFgSgshja9Y36Q+caHUzTDY6rrIy9HCSrjnbXmx3pCtvNZHmR8xg==",
   "_location": "/mongodb",
   "_phantomChildren": {},
   "_requested": {
     "type": "version",
     "registry": true,
-    "raw": "mongodb@3.6.8",
+    "raw": "mongodb@4.1.2",
     "name": "mongodb",
     "escapedName": "mongodb",
-    "rawSpec": "3.6.8",
+    "rawSpec": "4.1.2",
     "saveSpec": null,
-    "fetchSpec": "3.6.8"
+    "fetchSpec": "4.1.2"
   },
   "_requiredBy": [
     "/mongoose"
   ],
-  "_resolved": "https://registry.npmjs.org/mongodb/-/mongodb-3.6.8.tgz",
-  "_shasum": "3e2632af81915b3ff99b7681121ca0895e8ed407",
-  "_spec": "mongodb@3.6.8",
-  "_where": "D:\\Apache\\Apache24\\htdocs\\Server\\React\\DemoProject\\demo\\node_modules\\mongoose",
+  "_resolved": "https://registry.npmjs.org/mongodb/-/mongodb-4.1.2.tgz",
+  "_shasum": "36ab494db3a9a827df41ccb0d9b36a94bfeae8d7",
+  "_spec": "mongodb@4.1.2",
+  "_where": "D:\\Apache\\Apache24\\htdocs\\Server\\React\\DemoProject\\frontend\\node_modules\\mongoose",
+  "author": {
+    "name": "The MongoDB NodeJS Team",
+    "email": "dbx-node@mongodb.com"
+  },
   "bugs": {
-    "url": "https://github.com/mongodb/node-mongodb-native/issues"
+    "url": "https://jira.mongodb.org/projects/NODE/issues/"
   },
   "bundleDependencies": false,
   "dependencies": {
-    "bl": "^2.2.1",
-    "bson": "^1.1.4",
-    "denque": "^1.4.1",
-    "optional-require": "^1.0.3",
-    "safe-buffer": "^5.1.2",
-    "saslprep": "^1.0.0"
+    "bson": "^4.5.2",
+    "denque": "^2.0.1",
+    "mongodb-connection-string-url": "^2.0.0",
+    "saslprep": "^1.0.3"
   },
   "deprecated": false,
   "description": "The official MongoDB driver for Node.js",
   "devDependencies": {
-    "@types/chai": "^4.2.16",
-    "@types/mocha": "^8.2.2",
-    "@types/node": "^14.14.37",
-    "array-includes": "^3.1.3",
-    "chai": "^4.1.1",
+    "@istanbuljs/nyc-config-typescript": "^1.0.1",
+    "@microsoft/api-extractor": "^7.18.7",
+    "@microsoft/tsdoc-config": "^0.15.2",
+    "@types/chai": "^4.2.21",
+    "@types/chai-subset": "^1.3.3",
+    "@types/kerberos": "^1.1.1",
+    "@types/mocha": "^9.0.0",
+    "@types/node": "^16.7.13",
+    "@types/saslprep": "^1.0.1",
+    "@types/semver": "^7.3.8",
+    "@types/whatwg-url": "^8.2.1",
+    "@typescript-eslint/eslint-plugin": "^4.31.0",
+    "@typescript-eslint/parser": "^4.31.0",
+    "bluebird": "^3.7.2",
+    "chai": "^4.3.4",
     "chai-subset": "^1.6.0",
-    "chalk": "^2.4.2",
+    "chalk": "^4.1.2",
     "co": "4.6.0",
-    "eslint": "^7.10.0",
-    "eslint-config-prettier": "^6.11.0",
-    "eslint-plugin-es": "^3.0.1",
-    "eslint-plugin-prettier": "^3.1.3",
-    "jsdoc": "^3.5.5",
+    "downlevel-dts": "^0.7.0",
+    "eslint": "^7.32.0",
+    "eslint-config-prettier": "^8.3.0",
+    "eslint-plugin-jsdoc": "^36.1.0",
+    "eslint-plugin-prettier": "^4.0.0",
+    "eslint-plugin-tsdoc": "^0.2.14",
+    "js-yaml": "^4.1.0",
+    "jsdoc": "^3.6.7",
     "lodash.camelcase": "^4.3.0",
-    "mocha": "5.2.0",
-    "mocha-sinon": "^2.1.0",
-    "mongodb-extjson": "^2.1.1",
-    "mongodb-mock-server": "^1.0.1",
+    "mocha": "^9.1.1",
+    "mocha-sinon": "^2.1.2",
+    "mongodb-mock-server": "^2.0.1",
     "nyc": "^15.1.0",
-    "object.entries": "^1.1.3",
-    "prettier": "^1.19.1",
-    "semver": "^5.5.0",
-    "sinon": "^4.3.0",
-    "sinon-chai": "^3.2.0",
-    "snappy": "^6.3.4",
-    "spec-xunit-file": "0.0.1-3",
-    "standard-version": "^9.2.0",
-    "tslib": "^2.2.0",
-    "typescript": "^4.2.4",
-    "util.promisify": "^1.0.1",
-    "worker-farm": "^1.5.0",
-    "wtfnode": "^0.8.0",
-    "yargs": "^14.2.0"
+    "prettier": "^2.3.2",
+    "rimraf": "^3.0.2",
+    "semver": "^7.3.5",
+    "sinon": "^11.1.2",
+    "sinon-chai": "^3.7.0",
+    "source-map-support": "^0.5.19",
+    "standard-version": "^9.3.1",
+    "ts-node": "^10.2.1",
+    "tsd": "^0.17.0",
+    "typedoc": "^0.21.9",
+    "typescript": "4.3.5",
+    "typescript-cached-transpile": "^0.0.6",
+    "xml2js": "^0.4.23",
+    "yargs": "^17.1.1"
   },
   "engines": {
-    "node": ">=4"
+    "node": ">=12.9.0"
   },
   "files": [
-    "index.js",
-    "lib"
+    "lib",
+    "src",
+    "etc/prepare.js",
+    "mongodb.d.ts",
+    "mongodb.ts34.d.ts"
   ],
   "homepage": "https://github.com/mongodb/node-mongodb-native",
   "keywords": [
