@@ -89,7 +85,7 @@
     "lib": "./lib/mongoose"
   },
   "engines": {
-    "node": ">=4.0.0"
+    "node": ">=12.0.0"
   },
   "eslintConfig": {
     "extends": [
