@@ -1,27 +1,27 @@
 {
-  "_from": "mquery@3.2.5",
-  "_id": "mquery@3.2.5",
+  "_from": "mquery@4.0.0",
+  "_id": "mquery@4.0.0",
   "_inBundle": false,
-  "_integrity": "sha512-VjOKHHgU84wij7IUoZzFRU07IAxd5kWJaDmyUzQlbjHjyoeK5TNeeo8ZsFDtTYnSgpW6n/nMNIHvE3u8Lbrf4A==",
+  "_integrity": "sha512-nGjm89lHja+T/b8cybAby6H0YgA4qYC/lx6UlwvHGqvTq8bDaNeCwl1sY8uRELrNbVWJzIihxVd+vphGGn1vBw==",
   "_location": "/mquery",
   "_phantomChildren": {},
   "_requested": {
     "type": "version",
     "registry": true,
-    "raw": "mquery@3.2.5",
+    "raw": "mquery@4.0.0",
     "name": "mquery",
     "escapedName": "mquery",
-    "rawSpec": "3.2.5",
+    "rawSpec": "4.0.0",
     "saveSpec": null,
-    "fetchSpec": "3.2.5"
+    "fetchSpec": "4.0.0"
   },
   "_requiredBy": [
     "/mongoose"
   ],
-  "_resolved": "https://registry.npmjs.org/mquery/-/mquery-3.2.5.tgz",
-  "_shasum": "8f2305632e4bb197f68f60c0cffa21aaf4060c51",
-  "_spec": "mquery@3.2.5",
-  "_where": "D:\\Apache\\Apache24\\htdocs\\Server\\React\\DemoProject\\demo\\node_modules\\mongoose",
+  "_resolved": "https://registry.npmjs.org/mquery/-/mquery-4.0.0.tgz",
+  "_shasum": "6c62160ad25289e99e0840907757cdfd62bde775",
+  "_spec": "mquery@4.0.0",
+  "_where": "D:\\Apache\\Apache24\\htdocs\\Server\\React\\DemoProject\\frontend\\node_modules\\mongoose",
   "author": {
     "name": "Aaron Heckmann",
     "email": "aaron.heckmann+github@gmail.com"
@@ -31,64 +31,20 @@
   },
   "bundleDependencies": false,
   "dependencies": {
-    "bluebird": "3.5.1",
-    "debug": "3.1.0",
+    "debug": "4.x",
     "regexp-clone": "^1.0.0",
-    "safe-buffer": "5.1.2",
     "sliced": "1.0.1"
   },
   "deprecated": false,
   "description": "Expressive query building for MongoDB",
   "devDependencies": {
     "eslint": "5.x",
-    "mocha": "4.1.0",
-    "mongodb": "3.6.1"
+    "eslint-plugin-mocha-no-only": "1.1.0",
+    "mocha": "9.x",
+    "mongodb": "4.x"
   },
   "engines": {
-    "node": ">=4.0.0"
-  },
-  "eslintConfig": {
-    "env": {
-      "node": true,
-      "mocha": true,
-      "es6": false
-    },
-    "extends": "eslint:recommended",
-    "parserOptions": {
-      "ecmaVersion": 5
-    },
-    "rules": {
-      "comma-style": "error",
-      "consistent-this": [
-        "error",
-        "_this"
-      ],
-      "indent": [
-        "error",
-        2,
-        {
-          "SwitchCase": 1,
-          "VariableDeclarator": 2
-        }
-      ],
-      "keyword-spacing": "error",
-      "no-console": "off",
-      "no-multi-spaces": "error",
-      "func-call-spacing": "error",
-      "no-trailing-spaces": "error",
-      "quotes": [
-        "error",
-        "single"
-      ],
-      "semi": "error",
-      "space-before-blocks": "error",
-      "space-before-function-paren": [
-        "error",
-        "never"
-      ],
-      "space-infix-ops": "error",
-      "space-unary-ops": "error"
-    }
+    "node": ">=12.0.0"
   },
   "homepage": "https://github.com/aheckmann/mquery/",
   "keywords": [
@@ -108,5 +64,5 @@
     "lint": "eslint .",
     "test": "mocha test/index.js test/*.test.js"
   },
-  "version": "3.2.5"
+  "version": "4.0.0"
 }
