@@ -1,27 +1,27 @@
 {
-  "_from": "nth-check@^2.0.0",
-  "_id": "nth-check@2.0.0",
+  "_from": "nth-check@2.0.1",
+  "_id": "nth-check@2.0.1",
   "_inBundle": false,
-  "_integrity": "sha512-i4sc/Kj8htBrAiH1viZ0TgU8Y5XqCaV/FziYK6TBczxmeKm3AEFWqqF3195yKudrarqy7Zu80Ra5dobFjn9X/Q==",
+  "_integrity": "sha512-it1vE95zF6dTT9lBsYbxvqh0Soy4SPowchj0UBGj/V6cTPnXXtQOPUbhZ6CmGzAD/rW22LQK6E96pcdJXk4A4w==",
   "_location": "/renderkid/nth-check",
   "_phantomChildren": {},
   "_requested": {
-    "type": "range",
+    "type": "version",
     "registry": true,
-    "raw": "nth-check@^2.0.0",
+    "raw": "nth-check@2.0.1",
     "name": "nth-check",
     "escapedName": "nth-check",
-    "rawSpec": "^2.0.0",
+    "rawSpec": "2.0.1",
     "saveSpec": null,
-    "fetchSpec": "^2.0.0"
+    "fetchSpec": "2.0.1"
   },
   "_requiredBy": [
     "/renderkid/css-select"
   ],
-  "_resolved": "https://registry.npmjs.org/nth-check/-/nth-check-2.0.0.tgz",
-  "_shasum": "1bb4f6dac70072fc313e8c9cd1417b5074c0a125",
-  "_spec": "nth-check@^2.0.0",
-  "_where": "D:\\Apache\\Apache24\\htdocs\\Server\\React\\DemoProject\\demo\\node_modules\\renderkid\\node_modules\\css-select",
+  "_resolved": "https://registry.npmjs.org/nth-check/-/nth-check-2.0.1.tgz",
+  "_shasum": "2efe162f5c3da06a28959fbd3db75dbeea9f0fc2",
+  "_spec": "nth-check@2.0.1",
+  "_where": "D:\\Apache\\Apache24\\htdocs\\Server\\React\\DemoProject\\frontend\\node_modules\\renderkid\\node_modules\\css-select",
   "author": {
     "name": "Felix Boehm",
     "email": "me@feedic.com"
@@ -36,16 +36,16 @@
   "deprecated": false,
   "description": "Parses and compiles CSS nth-checks to highly optimized functions.",
   "devDependencies": {
-    "@types/jest": "^26.0.0",
-    "@types/node": "^14.0.5",
-    "@typescript-eslint/eslint-plugin": "^4.1.0",
-    "@typescript-eslint/parser": "^4.1.0",
-    "eslint": "^7.0.0",
-    "eslint-config-prettier": "^6.0.0",
-    "jest": "^26.0.1",
-    "prettier": "^2.1.1",
-    "ts-jest": "^26.0.0",
-    "typescript": "^4.0.2"
+    "@types/jest": "^27.0.1",
+    "@types/node": "^16.9.1",
+    "@typescript-eslint/eslint-plugin": "^4.31.1",
+    "@typescript-eslint/parser": "^4.31.1",
+    "eslint": "^7.32.0",
+    "eslint-config-prettier": "^8.3.0",
+    "jest": "^27.2.0",
+    "prettier": "^2.4.1",
+    "ts-jest": "^27.0.5",
+    "typescript": "^4.4.3"
   },
   "directories": {
     "lib": "lib/"
@@ -78,7 +78,6 @@
   },
   "scripts": {
     "build": "tsc",
-    "coverage": "cat coverage/lcov.info | coveralls",
     "format": "npm run format:es && npm run format:prettier",
     "format:es": "npm run lint:es -- --fix",
     "format:prettier": "npm run prettier -- --write",
@@ -87,9 +86,10 @@
     "lint:prettier": "npm run prettier -- --check",
     "prepare": "npm run build",
     "prettier": "prettier '**/*.{ts,md,json,yml}'",
-    "test": "jest --coverage && npm run lint"
+    "test": "npm run test:jest && npm run lint",
+    "test:jest": "jest"
   },
   "sideEffects": false,
   "types": "lib/index.d.ts",
-  "version": "2.0.0"
+  "version": "2.0.1"
 }
