@@ -8,29 +8,31 @@ const WhoCanContactUsSchema = new Schema({
     tenant_type: {
         type: String,
         trim: true,
+        unique: true,
         required: [true, 'Tenant type is required.'],
         minLength: [3, 'Tenant type must have minimum 3 characters.'],
         maxLength: [100, 'Maximum character for tenant type should be 100 characters.']
     },
+    status: {
+        type: Number,
+        trim: true,
+    },
     created_by: {
         type: 'ObjectId',
-        required: [true, "User ID is required."],
         ref: 'Users'
     },
     modified_by: {
         type: 'ObjectId',
-        required: [true, "User ID is required"],
         ref: 'Users'
     },
     created_on: {
-        type: Date,
-        default: Date.now
+        type: Date
     },
     modified_on: {
-        type: Date,
-        default: Date.now
+        type: Date
     }
 });
 
 
-module.exports = mongoose.model("WhoCanContactUs", WhoCanContactUsSchema);
+const WhoCanContactUs = mongoose.model("WhoCanContactUs", WhoCanContactUsSchema);
+export default WhoCanContactUs;
