@@ -5,20 +5,25 @@ import Users from '../models/UserModel.js';
 import { logger } from '../config/configLoggers.js'
 
 
-const verifyToken = asyncHandler(async(req, res, next) => {
+export const verifyToken = asyncHandler(async(req, res, next) => {
     let token;
     if (req.headers.authorization && req.headers.authorization.startsWith('Bearer')) {
         try {
             token = req.headers.authorization.split(' ')[1];
             const decoded = jwt.verify(token, process.env.JWT_SECRET);
-            console.log(decoded);
             req.user = await Users.findById(decoded.user_id).select('-password');
             next();
         } catch (error) {
             logger.error(error);
             //console.error(error)
             res.status(401)
-            throw new Error('Not authorized, token failed');
+            throw new Error('Token expired.');
+            /* if(error instanceof jwt.TokenExpiredError) {
+                res.status(401).json({status: 'error', message: 'TokenExpiredError'});
+            } else {
+                res.status(401)
+                throw new Error('Token expired.');
+            } */
         }
     }
     
@@ -28,4 +33,22 @@ const verifyToken = asyncHandler(async(req, res, next) => {
     }
 })
 
-export { verifyToken }
+export const generateToken = (user_id) => {
+    return jwt.sign({ user_id:  user_id}, process.env.JWT_SECRET, { expiresIn: '8h' });
+};
+
+export const getUserId = async(req) => {
+    let token;
+    if (req.headers.authorization && req.headers.authorization.startsWith('Bearer')) {
+        try {
+            token = req.headers.authorization.split(' ')[1];
+            const decoded = await jwt.verify(token, process.env.JWT_SECRET);
+            return decoded.user_id;
+        } catch (error) {
+            return false;
+        }
+    }
+    if (!token) {
+        return false;
+    }
+}
