@@ -1,5 +1,5 @@
+import jwt from 'jsonwebtoken';
 import mongoose from 'mongoose';
-import  jwt from 'jsonwebtoken';
 import { logger } from '../config/configLoggers.js'
 
 const notFound = (req, res, next) => {
@@ -10,7 +10,7 @@ const notFound = (req, res, next) => {
 
 const errorHandler = (err, req, res, next) => {
     let CUSTOM_ERRORS = ['Not authorized, token failed', 'Not authorized, no token'];
-    console.log('Hanler==>'+err);
+    console.log('Error Middleware==>'+err);
     let statusCode = res.statusCode === 200 ? 500 : res.statusCode;
     if(CUSTOM_ERRORS.indexOf(err.message) != -1) {
         statusCode = 200;
