@@ -7,6 +7,6 @@ import { verifyToken } from '../middleware/authMiddleware.js'
 const { checkSchema } = validator;
 const router = express.Router()
 
-router.route('/asset/add').post(verifyToken, checkSchema(assetValidationSchema), createAsset)
+router.route('/add').post(verifyToken, checkSchema(assetValidationSchema), createAsset)
 
 export default router;
