@@ -1,14 +1,15 @@
 import React from 'react';
 import Select from 'react-select';
 import CreatableSelect from 'react-select/creatable';
-import { useSelector, useDispatch, connect } from 'react-redux';
+import { connect } from 'react-redux';
+import { toast } from 'react-toastify';
 import { useLocation, useHistory } from 'react-router';
 import { CUSTOM_STYLES, BULK_DATA_INSERT_TYPES } from '../../../config/globalConstant';
 import LoadingIndicator from '../common/LoadingIndicator';
+import Loader from '../../Loader';
 import { getAllOptions } from '../../../actions/assetActions';
-import { bulkInsert } from '../../../actions/settingActions';
+import { bulkInsert, createSingleOption, dynamicOptionMerge, resetResult } from '../../../actions/settingActions';
 import { bindActionCreators } from 'redux';
-
 
 
 class BulkDataInsert extends React.Component {
@@ -21,14 +22,37 @@ class BulkDataInsert extends React.Component {
             default_type_option: 0,
             bulk_data: "",
             error_msg: "",
-            bulk_loader:false
+            bulk_loader:false,
+            loader: false
         }
 
         this.onSelectDataType.bind(this);
         this.onSelectOptions.bind(this);
         this.handleBulkData.bind(this);
         this.insertBulkData.bind(this);
+        this.toUniqueArray.bind(this);
         
+    }
+    componentDidUpdate() {
+        if(this.props.settingState.bulk_result != undefined && Object.keys(this.props.settingState.bulk_result).length != 0) {
+            if(this.props.settingState.bulk_result.status == 'error') {
+                toast.error(this.props.settingState.bulk_result.message, {theme: "colored"});
+            } else {
+                if(this.props.settingState.bulk_result.result != undefined) {
+                    let pushValue = {
+                        option: {
+                            label: this.props.settingState.bulk_result.result.city_name,
+                            value: this.props.settingState.bulk_result.result._id
+                        },
+                        fieldType: this.state.type_of_data_to_insert
+                    };
+                    this.props.dynamicOptionMerge(pushValue);
+                    this.setState({default_type_option: pushValue.option});
+                }
+                toast.success(this.props.settingState.bulk_result.message, {theme: "colored"});
+            }
+            this.props.resetResult();
+        }
     }
     componentDidMount() {
         this.props.getoptions();
@@ -51,12 +75,14 @@ class BulkDataInsert extends React.Component {
     onSelectOptions(value, actionMeta) {
         console.log(value, actionMeta);
         if(actionMeta.action == "create-option" && actionMeta.option.__isNew__) {
-            //this.setState({default_type_option: })
+            let option = value.value;
+            this.props.createSingleOption(this.state.type_of_data_to_insert, option, this.state);
+        } else {
+            this.setState({default_type_option: value});
         }
     };
     handleBulkData(e) {
         this.setState({error_msg: ''});
-        console.log(e.target.value);
         this.setState({bulk_data: e.target.value});
     }
     insertBulkData(e) {
@@ -65,51 +91,74 @@ class BulkDataInsert extends React.Component {
             this.setState({error_msg: 'Data Required'});
         } else {
             this.setState({error_msg: ''});
-            const bulkData = this.state.bulk_data.split(",");
-            console.log(bulkData);
+            let bulkData = this.state.bulk_data;
+            bulkData = bulkData.trim();
+            bulkData = bulkData.split(",");
+            bulkData = this.toUniqueArray(bulkData);
+            setTimeout(() => {
+                this.props.bulkInsert(this.state.type_of_data_to_insert, bulkData, this.state);
+            }, 50)
         }
+    }
+    toUniqueArray(a){
+        var newArr = [];
+        for (var i = 0; i < a.length; i++) {
+            if (newArr.indexOf(a[i].trim()) === -1 && a[i].trim() != "") {
+                newArr.push(a[i].trim());
+            }
+        }
+      return newArr;
     }
     render() {
         
         return (
             <section id="BulkDataInsert">
-                <h4 className="mt-4 pt-5">
-                    Bulk Data Insert
-                </h4>
-                <form id="bulkDataForm" name="bulkDataForm" onSubmit={e => this.insertBulkData(e)}>
-                <div className="my-2">
-                    <div className="my-2">
-                        Select Data want to insert:
-                    </div>
-                    <Select isLoading={this.state.bulk_loader} className="basic-single" classNamePrefix="select" id="type_of_data_to_insert" name="type_of_data_to_insert" options={BULK_DATA_INSERT_TYPES} onChange={e => this.onSelectDataType(e)} defaultValue={this.state.type_of_data_to_insert} components={{LoadingIndicator}} styles={CUSTOM_STYLES} />
-                    <span id="type_of_data_to_insert_error"></span>
-                </div>
-                {this.state.type_of_data_to_insert !=0 && 
-                    <div>
+                {this.state.loader == true ? <Loader /> : <div>
+                        <h4 className="mt-4 pt-5">
+                            Bulk Data Insert
+                        </h4>
+                        <div>
+                            {JSON.stringify(this.state)}
+                        </div>
+                        <form id="bulkDataForm" name="bulkDataForm" onSubmit={e => this.insertBulkData(e)}>
                         <div className="my-2">
                             <div className="my-2">
-                                {BULK_DATA_INSERT_TYPES[(this.state.type_of_data_to_insert - 1)].label}
+                                Select Data want to insert:
                             </div>
-                            <CreatableSelect isLoading={this.state.bulk_loader} formatCreateLabel={(inputValue) => 'Add or Create "'+inputValue+'"'} createOptionPosition="first" className="basic-single" classNamePrefix="select" id="who_can_contact" name="who_can_contact" options={this.state.type_options} createOptionPosition={"first"} onChange={(value, actionMeta) => this.onSelectOptions(value, actionMeta)} defaultValue={this.state.default_type_option} components={{LoadingIndicator}} styles={CUSTOM_STYLES} />
-                            <span></span>
+                            <Select isLoading={this.state.bulk_loader} className="basic-single" classNamePrefix="select" id="type_of_data_to_insert" name="type_of_data_to_insert" options={BULK_DATA_INSERT_TYPES} onChange={e => this.onSelectDataType(e)} defaultValue={this.state.type_of_data_to_insert} components={{LoadingIndicator}} styles={CUSTOM_STYLES} />
+                            <span id="type_of_data_to_insert_error"></span>
                         </div>
-                        <div className="mb-3">
-                            <label htmlFor="bulk_data" className="form-label">Bulk Data</label>
-                            <textarea className="form-control" id="bulk_data" name="bulk_data" rows="3" onChange={e => this.handleBulkData(e)} value={this.bulk_data}></textarea>
-                            <span id="bulk_data_error" className="text-danger">{this.state.error_msg}</span>
-                        </div>
-                        <div>
-                            <button className="btn btn-primary">
-                                Submit
-                            </button>
-                        </div>
+                        {this.state.type_of_data_to_insert !=0 && 
+                            <div>
+                                <div className="my-2">
+                                    <div className="my-2">
+                                        {BULK_DATA_INSERT_TYPES[(this.state.type_of_data_to_insert - 1)].label}
+                                    </div>
+                                    <CreatableSelect isLoading={this.state.bulk_loader} formatCreateLabel={(inputValue) => 'Add or Create "'+inputValue+'"'} createOptionPosition="first" className="basic-single" classNamePrefix="select" id="who_can_contact" name="who_can_contact" options={this.state.type_options} createOptionPosition={"first"} onChange={(value, actionMeta) => this.onSelectOptions(value, actionMeta)} /* defaultValue={this.state.default_type_option} */ value={this.state.default_type_option} components={{LoadingIndicator}} styles={CUSTOM_STYLES} />
+                                    <span></span>
+                                </div>
+                                <div className="mb-3">
+                                    <label htmlFor="bulk_data" className="form-label">Bulk Data</label>
+                                    <textarea className="form-control" id="bulk_data" name="bulk_data" rows="3" onChange={e => this.handleBulkData(e)} value={this.bulk_data}></textarea>
+                                    <span id="bulk_data_error" className="text-danger">{this.state.error_msg}</span>
+                                </div>
+                                <div>
+                                    <button className="btn btn-primary">
+                                        Submit
+                                    </button>
+                                </div>
+                            </div>
+                        }
+                        </form>
+                        
                     </div>
                 }
-                </form>
             </section>
         )
     }
 }
+
+
 
 const mapStateToProps = (state) => {
     return {
@@ -123,8 +172,17 @@ const mapDispatchToProps = dispatch => {
         getoptions: () => {
             dispatch(getAllOptions())
         },
-        bulkInsert: () => {
-            dispatch(bulkInsert())
+        bulkInsert: (field, bulkValues, state) => {
+            dispatch(bulkInsert(field, bulkValues, state))
+        },
+        createSingleOption: (field, option, state) => {
+            dispatch(createSingleOption(field, option, state))
+        },
+        dynamicOptionMerge: (pushValue) => {
+            dispatch(dynamicOptionMerge(pushValue))
+        },
+        resetResult: () => {
+            dispatch(resetResult())
         }
     }
 }
