@@ -10,13 +10,20 @@ import { Provider } from 'react-redux';
 import store from './store/store';
 
 ReactDOM.render(
+  <Provider store={store}>
+    <App />
+  </Provider>,
+  document.getElementById('root')
+);
+
+/* ReactDOM.render(
   <React.StrictMode>
       <Provider store={store}>
         <App />
       </Provider>
   </React.StrictMode>,
   document.getElementById('root')
-);
+); */
 
 // If you want to start measuring performance in your app, pass a function
 // to log results (for example: reportWebVitals(console.log))
